requirements.txt
plotly==4.8.2
dash==1.13.4
stockstats==0.3.1
dash_bootstrap_components==0.10.3
pandas==1.0.5
requests==2.24.0 
requests-html==0.10.0 
yahoo_fin==0.8.6
dash_html_components==1.0.3
dash_table==4.8.1
dash_core_components==1.10.1
numpy==1.19.0
gunicorn==19.7.1
yfinance==0.1.59

cachetools==4.2.1
mysql-connector-python==8.0.23
psycopg2==2.8.6
python-dotenv==0.16.0
sqlalchemy==1.3.23
praw==7.2.0
Unidecode==1.2.0